Moved the onboarding into new components
