+++ apps/webapp/app/components/StepContentContainer.tsx
@@ -1,0 +1,3 @@
+export function StepContentContainer({ children }: { children: React.ReactNode }) {
+  return <div className="mb-6 ml-9 mt-1">{children}</div>;
+}
