+++ apps/webapp/app/components/helpContent/HelpContentText.tsx
@@ -1,268 +1,26 @@
-import { ChatBubbleLeftRightIcon } from "@heroicons/react/20/solid";
-import { Link, useRevalidator } from "@remix-run/react";
-import { useEffect, useState } from "react";
-import { useEventSource } from "remix-utils";
-import invariant from "tiny-invariant";
-import gradientBackground from "~/assets/images/gradient-background.png";
+import { Link } from "@remix-run/react";
 import { Paragraph } from "~/components/primitives/Paragraph";
 import { StepNumber } from "~/components/primitives/StepNumber";
-import { useAppOrigin } from "~/hooks/useAppOrigin";
-import { useDevEnvironment } from "~/hooks/useEnvironments";
 import { useJob } from "~/hooks/useJob";
 import { useOrganization } from "~/hooks/useOrganizations";
 import { useProject } from "~/hooks/useProject";
 import { IntegrationIcon } from "~/routes/_app.orgs.$organizationSlug.projects.$projectParam.integrations/route";
-import { jobTestPath, projectStreamingPath } from "~/utils/pathBuilder";
-import { Feedback } from "../Feedback";
+import { jobTestPath } from "~/utils/pathBuilder";
 import { CodeBlock } from "../code/CodeBlock";
 import { InlineCode } from "../code/InlineCode";
 import { EnvironmentLabel } from "../environments/EnvironmentLabel";
 import { HelpPanelProps } from "../integrations/ApiKeyHelp";
 import { HelpInstall } from "../integrations/HelpInstall";
 import { HelpSamples } from "../integrations/HelpSamples";
-import { Button, LinkButton } from "../primitives/Buttons";
+import { LinkButton } from "../primitives/Buttons";
 import { Callout, variantClasses } from "../primitives/Callout";
-import {
-  ClientTabs,
-  ClientTabsContent,
-  ClientTabsList,
-  ClientTabsTrigger,
-} from "../primitives/ClientTabs";
-import { ClipboardField } from "../primitives/ClipboardField";
-import { Header1, Header2 } from "../primitives/Headers";
-import { NamedIcon } from "../primitives/NamedIcon";
-import { RadioGroup, RadioGroupItem } from "../primitives/RadioButton";
+import { Header2 } from "../primitives/Headers";
 import { TextLink } from "../primitives/TextLink";
 import integrationButton from "./integration-button.png";
 import selectEnvironment from "./select-environment.png";
 import selectExample from "./select-example.png";
-
-type SelectionChoices = "use-existing-project" | "create-new-next-app";
-
-export function HowToSetupYourProject() {
-  const project = useProject();
-  const devEnvironment = useDevEnvironment();
-  const appOrigin = useAppOrigin();
-
-  const [selectedValue, setSelectedValue] = useState<SelectionChoices | null>(null);
-
-  invariant(devEnvironment, "devEnvironment is required");
-
-  const revalidator = useRevalidator();
-  const events = useEventSource(projectStreamingPath(project.id), {
-    event: "message",
-  });
-
-  useEffect(() => {
-    if (events !== null) {
-      // This uses https://www.npmjs.com/package/canvas-confetti
-      if ("confetti" in window && typeof window.confetti !== "undefined") {
-        const duration = 3.5 * 1000;
-        const animationEnd = Date.now() + duration;
-        const defaults = {
-          startVelocity: 30,
-          spread: 360,
-          ticks: 60,
-          zIndex: 0,
-          colors: [
-            "#E7FF52",
-            "#41FF54",
-            "rgb(245 158 11)",
-            "rgb(22 163 74)",
-            "rgb(37 99 235)",
-            "rgb(67 56 202)",
-            "rgb(219 39 119)",
-            "rgb(225 29 72)",
-            "rgb(217 70 239)",
-          ],
-        };
-        function randomInRange(min: number, max: number): number {
-          return Math.random() * (max - min) + min;
-        }
-        // @ts-ignore
-        const interval = setInterval(function () {
-          const timeLeft = animationEnd - Date.now();
-
-          if (timeLeft <= 0) {
-            return clearInterval(interval);
-          }
-
-          const particleCount = 60 * (timeLeft / duration);
-          // since particles fall down, start a bit higher than random
-          // @ts-ignore
-          window.confetti(
-            Object.assign({}, defaults, {
-              particleCount,
-              origin: { x: randomInRange(0.1, 0.4), y: Math.random() - 0.2 },
-            })
-          );
-          // @ts-ignore
-          window.confetti(
-            Object.assign({}, defaults, {
-              particleCount,
-              origin: { x: randomInRange(0.6, 0.9), y: Math.random() - 0.2 },
-            })
-          );
-        }, 250);
-      }
-
-      revalidator.revalidate();
-    }
-    // WARNING Don't put the revalidator in the useEffect deps array or bad things will happen
-  }, [events]); // eslint-disable-line react-hooks/exhaustive-deps
-
-  return (
-    <div
-      className="-ml-4 -mt-4 h-full w-[calc(100%+32px)] bg-cover bg-no-repeat pt-20"
-      style={{ backgroundImage: `url("${gradientBackground}")` }}
-    >
-      <div className="mx-auto max-w-3xl">
-        <div className="flex items-center justify-between">
-          <Header1 spacing className="text-bright">
-            Get setup in {selectedValue === "create-new-next-app" ? "5" : "2"} minutes
-          </Header1>
-          <Feedback
-            button={
-              <Button variant="secondary/small" LeadingIcon={ChatBubbleLeftRightIcon}>
-                I'm stuck!
-              </Button>
-            }
-            defaultValue="help"
-          />
-        </div>
-        <RadioGroup
-          className="mb-4 flex gap-x-2"
-          onValueChange={(value) => setSelectedValue(value as SelectionChoices)}
-        >
-          <RadioGroupItem
-            label="Use an existing Next.js project"
-            description="Use Trigger.dev in an existing Next.js project in less than 2 mins."
-            value="use-existing-project"
-            checked={selectedValue === "use-existing-project"}
-            variant="icon"
-            data-action="use-existing-project"
-            icon={<NamedIcon className="h-12 w-12 text-green-600" name={"tree"} />}
-          />
-          <RadioGroupItem
-            label="Create a new Next.js project"
-            description="This is the quickest way to try out Trigger.dev in a new Next.js project and takes 5 mins."
-            value="create-new-next-app"
-            checked={selectedValue === "create-new-next-app"}
-            variant="icon"
-            data-action="create-new-next-app"
-            icon={<NamedIcon className="h-8 w-8 text-green-600" name={"sapling"} />}
-          />
-        </RadioGroup>
-        {selectedValue && (
-          <>
-            {selectedValue === "create-new-next-app" ? (
-              <>
-                <StepNumber stepNumber="1" title="Create a new Next.js project" />
-                <StepContentContainer>
-                  <ClientTabs defaultValue="npm">
-                    <ClientTabsList>
-                      <ClientTabsTrigger value={"npm"}>npm</ClientTabsTrigger>
-                      <ClientTabsTrigger value={"pnpm"}>pnpm</ClientTabsTrigger>
-                      <ClientTabsTrigger value={"yarn"}>yarn</ClientTabsTrigger>
-                    </ClientTabsList>
-                    <ClientTabsContent value={"npm"}>
-                      <ClipboardField
-                        variant="primary/medium"
-                        className="mb-4"
-                        value={`npx create-next-app@latest`}
-                      />
-                    </ClientTabsContent>
-                    <ClientTabsContent value={"pnpm"}>
-                      <ClipboardField
-                        variant="primary/medium"
-                        className="mb-4"
-                        value={`pnpm create next-app`}
-                      />
-                    </ClientTabsContent>
-                    <ClientTabsContent value={"yarn"}>
-                      <ClipboardField
-                        variant="primary/medium"
-                        className="mb-4"
-                        value={`yarn create next-app`}
-                      />
-                    </ClientTabsContent>
-                  </ClientTabs>
-
-                  <Paragraph spacing variant="small">
-                    Trigger.dev works with either the Pages or App Router configuration.
-                  </Paragraph>
-                </StepContentContainer>
-                <StepNumber stepNumber="2" title="Navigate to your new Next.js project" />
-                <StepContentContainer>
-                  <Paragraph spacing>
-                    You have now created a new Next.js project. Let’s <InlineCode>cd</InlineCode>{" "}
-                    into it using the project name you just provided:
-                  </Paragraph>
-                  <ClipboardField
-                    value={"cd [replace with your project name]"}
-                    variant={"primary/medium"}
-                  ></ClipboardField>
-                </StepContentContainer>
-                <StepNumber
-                  stepNumber="3"
-                  title="Run the CLI 'init' command in your new Next.js project"
-                />
-                <StepContentContainer>
-                  <InitCommand appOrigin={appOrigin} apiKey={devEnvironment.apiKey} />
-                  <Paragraph spacing variant="small">
-                    You’ll notice a new folder in your project called 'jobs'. We’ve added a very
-                    simple example Job in <InlineCode variant="extra-small">examples.ts</InlineCode>{" "}
-                    to help you get started.
-                  </Paragraph>
-                </StepContentContainer>
-                <StepNumber stepNumber="4" title="Run your Next.js app" />
-                <StepContentContainer>
-                  <NextDevCommand />
-                </StepContentContainer>
-                <StepNumber stepNumber="5" title="Run the CLI 'dev' command" />
-                <StepContentContainer>
-                  <TriggerDevStep />
-                </StepContentContainer>
-                <StepNumber stepNumber="6" title="Wait for Jobs" displaySpinner />
-                <StepContentContainer>
-                  <Paragraph>This page will automatically refresh.</Paragraph>
-                </StepContentContainer>
-              </>
-            ) : (
-              <>
-                <StepNumber
-                  stepNumber="1"
-                  title="Run the CLI 'init' command in an existing Next.js project"
-                />
-                <StepContentContainer>
-                  <InitCommand appOrigin={appOrigin} apiKey={devEnvironment.apiKey} />
-
-                  <Paragraph spacing variant="small">
-                    You’ll notice a new folder in your project called 'jobs'. We’ve added a very
-                    simple example Job in <InlineCode variant="extra-small">examples.ts</InlineCode>{" "}
-                    to help you get started.
-                  </Paragraph>
-                </StepContentContainer>
-                <StepNumber stepNumber="2" title="Run your Next.js app" />
-                <StepContentContainer>
-                  <NextDevCommand />
-                </StepContentContainer>
-                <StepNumber stepNumber="3" title="Run the CLI 'dev' command" />
-                <StepContentContainer>
-                  <TriggerDevStep />
-                </StepContentContainer>
-                <StepNumber stepNumber="4" title="Wait for Jobs" displaySpinner />
-                <StepContentContainer>
-                  <Paragraph>This page will automatically refresh.</Paragraph>
-                </StepContentContainer>
-              </>
-            )}
-          </>
-        )}
-      </div>
-    </div>
-  );
-}
+import { StepContentContainer } from "../StepContentContainer";
+import { TriggerDevCommand } from "../SetupCommands";
 
 export function HowToRunYourJob() {
   const organization = useOrganization();
@@ -550,119 +308,6 @@ export function HowToUseApiKeysAndEndpoints() {
       <Callout variant="docs" to="https://trigger.dev/docs/documentation/guides/deployment">
         Read the deployment guide to learn more.
       </Callout>
-    </>
-  );
-}
-
-function StepContentContainer({ children }: { children: React.ReactNode }) {
-  return <div className="mb-6 ml-9 mt-1">{children}</div>;
-}
-
-function InitCommand({ appOrigin, apiKey }: { appOrigin: string; apiKey: string }) {
-  return (
-    <ClientTabs defaultValue="npm">
-      <ClientTabsList>
-        <ClientTabsTrigger value={"npm"}>npm</ClientTabsTrigger>
-        <ClientTabsTrigger value={"pnpm"}>pnpm</ClientTabsTrigger>
-        <ClientTabsTrigger value={"yarn"}>yarn</ClientTabsTrigger>
-      </ClientTabsList>
-      <ClientTabsContent value={"npm"}>
-        <ClipboardField
-          variant="primary/medium"
-          className="mb-4"
-          secure={`npx @trigger.dev/cli@latest init -k ••••••••• -t ${appOrigin}`}
-          value={`npx @trigger.dev/cli@latest init -k ${apiKey} -t ${appOrigin}`}
-        />
-      </ClientTabsContent>
-      <ClientTabsContent value={"pnpm"}>
-        <ClipboardField
-          variant="primary/medium"
-          className="mb-4"
-          secure={`pnpm dlx @trigger.dev/cli@latest init -k ••••••••• -t ${appOrigin}`}
-          value={`pnpm dlx @trigger.dev/cli@latest init -k ${apiKey} -t ${appOrigin}`}
-        />
-      </ClientTabsContent>
-      <ClientTabsContent value={"yarn"}>
-        <ClipboardField
-          variant="primary/medium"
-          className="mb-4"
-          secure={`yarn dlx @trigger.dev/cli@latest init -k ••••••••• -t ${appOrigin}`}
-          value={`yarn dlx @trigger.dev/cli@latest init -k ${apiKey} -t ${appOrigin}`}
-        />
-      </ClientTabsContent>
-    </ClientTabs>
-  );
-}
-
-function NextDevCommand() {
-  return (
-    <ClientTabs defaultValue="npm">
-      <ClientTabsList>
-        <ClientTabsTrigger value={"npm"}>npm</ClientTabsTrigger>
-        <ClientTabsTrigger value={"pnpm"}>pnpm</ClientTabsTrigger>
-        <ClientTabsTrigger value={"yarn"}>yarn</ClientTabsTrigger>
-      </ClientTabsList>
-      <ClientTabsContent value={"npm"}>
-        <ClipboardField variant="primary/medium" className="mb-4" value={`npm run dev`} />
-      </ClientTabsContent>
-      <ClientTabsContent value={"pnpm"}>
-        <ClipboardField variant="primary/medium" className="mb-4" value={`pnpm run dev`} />
-      </ClientTabsContent>
-      <ClientTabsContent value={"yarn"}>
-        <ClipboardField variant="primary/medium" className="mb-4" value={`yarn run dev`} />
-      </ClientTabsContent>
-    </ClientTabs>
-  );
-}
-
-function TriggerDevCommand() {
-  return (
-    <ClientTabs defaultValue="npm">
-      <ClientTabsList>
-        <ClientTabsTrigger value={"npm"}>npm</ClientTabsTrigger>
-        <ClientTabsTrigger value={"pnpm"}>pnpm</ClientTabsTrigger>
-        <ClientTabsTrigger value={"yarn"}>yarn</ClientTabsTrigger>
-      </ClientTabsList>
-      <ClientTabsContent value={"npm"}>
-        <ClipboardField
-          variant="primary/medium"
-          className="mb-4"
-          value={`npx @trigger.dev/cli@latest dev`}
-        />
-      </ClientTabsContent>
-      <ClientTabsContent value={"pnpm"}>
-        <ClipboardField
-          variant="primary/medium"
-          className="mb-4"
-          value={`pnpm dlx @trigger.dev/cli@latest dev`}
-        />
-      </ClientTabsContent>
-      <ClientTabsContent value={"yarn"}>
-        <ClipboardField
-          variant="primary/medium"
-          className="mb-4"
-          value={`yarn dlx @trigger.dev/cli@latest dev`}
-        />
-      </ClientTabsContent>
-    </ClientTabs>
-  );
-}
-
-function TriggerDevStep() {
-  return (
-    <>
-      <Paragraph spacing>
-        In a <span className="text-amber-400">separate terminal window or tab</span> run:
-      </Paragraph>
-      <TriggerDevCommand />
-      <Paragraph spacing variant="small">
-        If you’re not running on port 3000 you can specify the port by adding{" "}
-        <InlineCode variant="extra-small">--port 3001</InlineCode> to the end.
-      </Paragraph>
-      <Paragraph spacing variant="small">
-        You should leave the <InlineCode variant="extra-small">dev</InlineCode> command running when
-        you're developing.
-      </Paragraph>
     </>
   );
 }
